feat: implement updatefield
